pyproject.toml
[tool.poetry]
name = "flask-logging"
version = "0.1.0"
description = "Logging tools for flask"
authors = ["Alex Rudy <opensource@alexrudy.net>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
pyyaml = "^5.3.1"
blinker = "^1.4"

[tool.poetry.dev-dependencies]
flake8-black = "^0.2.0"
black = "^19.10b0"
mypy = "^0.782"
pytest = "^5.4.3"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
